Review: three more collect-continuously modes, the wasm log lines, Windows' native architecture
regress-174463162.js also skips dfg-eager, dfg-eager-no-cjit-validate and
ftl-eager: in a release build those pass --collectContinuously=true too (five
of the default modes do, not two).

WasmOperations.cpp: the dataLogLnIf lines in the two `access` lambdas divide by
static_cast<int>(sizeof(...)) like the returns below them. They are compiled
out (verbose is a constexpr false), so nothing changes until someone turns it
on.

webkitdirs.pm: determineNativeArchitecture() reads PROCESSOR_ARCHITEW6432 /
PROCESSOR_ARCHITECTURE on Windows, where there is no uname, so that
run-javascriptcore-tests --architecture arm64 on Windows on ARM passes the
"cannot run tests with arm64 on this machine" check. Not exercised: the Windows
lanes are not tested at the moment.

+++ Source/JavaScriptCore/wasm/WasmOperations.cpp
@@ -106,7 +106,7 @@ JSC_DEFINE_JIT_OPERATION(operationJSToWasmEntryWrapperBuildFrame, JSToWasmCallee
     }
 
     auto access = [sp, callFrame]<typename V>(auto* arr, int i) -> V* {
-        dataLogLnIf(WasmOperationsInternal::verbose, "fp[", (&reinterpret_cast<uint8_t*>(arr)[i / sizeof(uint8_t)] - reinterpret_cast<uint8_t*>(callFrame)), "] sp[", (&reinterpret_cast<uint8_t*>(arr)[i / sizeof(uint8_t)] - reinterpret_cast<uint8_t*>(sp)), "](", RawHex(reinterpret_cast<V*>(arr)[i / sizeof(V)]), ")");
+        dataLogLnIf(WasmOperationsInternal::verbose, "fp[", (&reinterpret_cast<uint8_t*>(arr)[i / static_cast<int>(sizeof(uint8_t))] - reinterpret_cast<uint8_t*>(callFrame)), "] sp[", (&reinterpret_cast<uint8_t*>(arr)[i / static_cast<int>(sizeof(uint8_t))] - reinterpret_cast<uint8_t*>(sp)), "](", RawHex(reinterpret_cast<V*>(arr)[i / static_cast<int>(sizeof(V))]), ")");
         return &reinterpret_cast<V*>(arr)[i / static_cast<int>(sizeof(V))];
     };
 
@@ -166,7 +166,7 @@ JSC_DEFINE_JIT_OPERATION(operationJSToWasmEntryWrapperBuildReturnFrame, EncodedJ
     const RTT& signature = callee->rtt();
 
     auto access = [sp, callFrame]<typename V>(auto* arr, int i) -> V* {
-        dataLogLnIf(WasmOperationsInternal::verbose, "fp[", (&reinterpret_cast<uint8_t*>(arr)[i / sizeof(uint8_t)] - reinterpret_cast<uint8_t*>(callFrame)), "] sp[", (&reinterpret_cast<uint8_t*>(arr)[i / sizeof(uint8_t)] - reinterpret_cast<uint8_t*>(sp)), "](", reinterpret_cast<V*>(arr)[i / sizeof(V)], ")");
+        dataLogLnIf(WasmOperationsInternal::verbose, "fp[", (&reinterpret_cast<uint8_t*>(arr)[i / static_cast<int>(sizeof(uint8_t))] - reinterpret_cast<uint8_t*>(callFrame)), "] sp[", (&reinterpret_cast<uint8_t*>(arr)[i / static_cast<int>(sizeof(uint8_t))] - reinterpret_cast<uint8_t*>(sp)), "](", reinterpret_cast<V*>(arr)[i / static_cast<int>(sizeof(V))], ")");
         return &reinterpret_cast<V*>(arr)[i / static_cast<int>(sizeof(V))];
     };
